fairk-NBN-classification-fairness-flip/requirements.txt
numpy
scikit-learn
pandas
omegaconf
matplotlib
scipy
statsmodels
pydantic
python-dotenv
ucimlrepo